Make addPresenceToQuery a shared util so other arrays can use it too
